Separate GUI and logic in startAnalysis
